Move the vm to its own package.
