+++ internal/runtime/vm/vm_test.go
@@ -1,7 +1,7 @@
 // Copyright 2011 Google Inc. All Rights Reserved.
 // This file is available under the Apache license.
 
-package runtime
+package vm
 
 import (
 	"context"
